# Conflicts:
#	conanfile.py
#	include/version.hpp
#	src/CMakeLists.txt
#	test/cqasm-utils.cpp

+++ src/v1x/cqasm-values.cpp
@@ -32,14 +32,14 @@ Value promote(const Value &value, const types::Type &type) {
     // Integers promote to real
     if (type->as_real()) {
         if (const auto &const_int = value->as_const_int()) {
-            retval = tree::make<values::ConstReal>(static_cast<ConstReal>(const_int->value));
+            retval = tree::make<values::ConstReal>(static_cast<ConstReal>(static_cast<double>(const_int->value)));
         }
     }
 
     // Integers and reals promote to complex
     if (type->as_complex()) {
         if (const auto &const_int = value->as_const_int()) {
-            retval = tree::make<values::ConstComplex>(static_cast<ConstComplex>(const_int->value));
+            retval = tree::make<values::ConstComplex>(static_cast<ConstComplex>(static_cast<double>(const_int->value)));
         } else if (auto const_real = value->as_const_real()) {
             retval = tree::make<values::ConstComplex>(static_cast<ConstComplex>(const_real->value));
         }
